% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomp_recomp.r
\name{decomp_recomp}
\alias{decomp_recomp}
\title{Decompose/Recompose}
\usage{
decomp_recomp(x, exclude, center = TRUE, scale = FALSE)
}
\arguments{
\item{x}{A matrix or ddmatrix of numeric values.}

\item{exclude}{A vector of positive integer values which are the principal components
whose variance contribution should be removed.}

\item{center}{Logical; determines if the matrix should be centered first.}

\item{scale}{Logical; determines if the matrix should be scaled first.}
}
\value{
A matrix or ddmatrix, matching the same type as the input \code{x}.
}
\description{
Exclude the variance contributed by the principal components
enumerated in the vector \code{exclude} from the data matrix.
}
\examples{
\dontrun{
## TODO add mpirun message boilerplate
library(pbdDMAT)
init.grid()

x <- ddmatrix("rnorm", nrow=100, ncol=10, bldim=c(2,2))
# Remove the contribution of PC 1's variance
decomp_recomp(x, 1)

finalize()
}

}
\references{
Inspired by a question from Ahmed Moustafa: \url{https://twitter.com/AhmedMoustafa/status/646310686725812224}
}
\author{
Drew Schmidt
}
