% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rpca.r
\name{rpca}
\alias{rpca}
\title{Random PCA}
\usage{
rpca(x, k = 1, q = 3, retx = TRUE, center = TRUE, scale = FALSE)
}
\arguments{
\item{x}{The input data matrix.}

\item{k}{The number of singular values and/or left/right singular vectors
to estimate.}

\item{q}{An integer exponent, say 1, 2, or 3.  See the paper for details.}

\item{retx}{Logical;  determines if the rotated data should be returned.}

\item{center,scale}{Logical; determines if the data should be centered/scaled first.}
}
\value{
An object of class \code{prcomp}.
}
\description{
Random PCA
}
\author{
Drew Schmidt
}
\keyword{PCA}
\keyword{SVD}

