% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdML-package.r
\docType{package}
\name{pbdML-package}
\alias{pbdML-package}
\title{Machine Learning}
\description{
TODO
}
\details{
\tabular{ll}{ 
   Package: \tab pbdML \cr 
   Type: \tab Package \cr 
   License: \tab BSD 2-clause \cr 
   LazyLoad: \tab yes \cr 
}

This package requires an MPI library (OpenMPI, MPICH2, or LAM/MPI).
}
\references{
Programming with Big Data in R Website: \url{http://r-pbd.org/}
}
\author{
Drew Schmidt \email{schmidt AT math.utk.edu}, George Ostrouchov, and Wei-Chen Chen.
}
\keyword{Package}
