% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rpca.r
\name{rpca}
\alias{rpca}
\title{Random PCA}
\usage{
rpca(x, k = 1, q = 3, retx = TRUE, center = TRUE, scale = FALSE)
}
\arguments{
\item{x}{The input data matrix.}

\item{k}{}

\item{q}{}

\item{retx}{}

\item{center,scale}{Logical; determines if the data should be centered/scaled first.}
}
\value{
An object of class \code{prcomp}.
}
\description{
Random PCA
}
\keyword{PCA}
\keyword{SVD}

