% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpca.r
\name{rpca}
\alias{rpca}
\title{Random PCA}
\usage{
rpca(x, k = 1, q = 3, retx = TRUE, center = TRUE, scale = FALSE)
}
\arguments{
\item{x}{The input data matrix.}

\item{k}{The number of singular values and/or left/right singular vectors
to estimate.}

\item{q}{An integer exponent, say 1, 2, or 3.  See the paper for details.}

\item{retx}{Logical;  determines if the rotated data should be returned.}

\item{center, scale}{Logical; determines if the data should be centered/scaled first.}
}
\value{
An object of class \code{prcomp}.
}
\description{
Random PCA
}
\examples{
\dontrun{
x <- matrix(rnorm(30), 10)

rpca(x)
}

}
\references{
Halko, Martinsson, and Tropp. 2011. Finding structure with 
randomness: probabilistic algorithms for constructing approximate
matrix decompositions. SIAM Review 53 217-288.

Duda, R. O., Hart, P. E., & Stork, D. G. (2012). Pattern classification,
chapter 10. John Wiley & Sons.
}
\author{
Drew Schmidt
}
\keyword{PCA}
\keyword{SVD}
