% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsvd.r
\name{rsvd}
\alias{rsvd}
\title{Random SVD}
\usage{
rsvd(x, k = 1, q = 3, retu = TRUE, retvt = TRUE)
}
\arguments{
\item{x}{The input data matrix.}

\item{k}{The number of singular values and/or left/right singular vectors
to estimate.}

\item{q}{An integer exponent, say 1, 2, or 3.  See the paper for details.}

\item{retu}{Logical; should the left singular vectors ("U") be returned?}

\item{retvt}{Logical; should the transposed right singular vectors ("VT") be returned?}
}
\value{
A list cotaining the singular values, and, if requested, the
left and/or right singular vectors.
}
\description{
Random SVD
}
\examples{
\dontrun{
x <- matrix(rnorm(30), 10)

rsvd(x)
}

}
\author{
George Ostrouchov and Drew Schmidt
}
\references{
Halko, Martinsson, and Tropp. 2011. Finding structure with 
randomness: probabilistic algorithms for constructing approximate
matrix decompositions. SIAM Review 53 217-288.
}
\keyword{PCA}
\keyword{SVD}

